+++ lib/util/http/auth/type.go
@@ -8,6 +8,7 @@ import (
 type Config struct {
 	OAuth     OAuthConfig     `json:"oauth" yaml:"oauth"`
 	BasicAuth BasicAuthConfig `json:"basic_auth" yaml:"basic_auth"`
+	JWT       JWTConfig       `json:"jwt" yaml:"jwt"`
 }
 
 // NewConfig creates a new Config with default values.
@@ -15,12 +16,16 @@ func NewConfig() Config {
 	return Config{
 		OAuth:     NewOAuthConfig(),
 		BasicAuth: NewBasicAuthConfig(),
+		JWT:       NewJWTConfig(),
 	}
 }
 
 // Sign method to sign an HTTP request for configured auth strategies.
 func (c Config) Sign(req *http.Request) error {
 	if err := c.OAuth.Sign(req); err != nil {
+		return err
+	}
+	if err := c.JWT.Sign(req); err != nil {
 		return err
 	}
 	return c.BasicAuth.Sign(req)
@@ -33,6 +38,7 @@ func (c Config) Sign(req *http.Request) error {
 type WebsocketConfig struct {
 	OAuth     OAuthConfig     `json:"oauth" yaml:"oauth"`
 	BasicAuth BasicAuthConfig `json:"basic_auth" yaml:"basic_auth"`
+	JWT       JWTConfig       `json:"jwt" yaml:"jwt"`
 }
 
 // NewWebsocketConfig creates a new WebsocketConfig with default values.
@@ -40,12 +46,16 @@ func NewWebsocketConfig() WebsocketConfig {
 	return WebsocketConfig{
 		OAuth:     NewOAuthConfig(),
 		BasicAuth: NewBasicAuthConfig(),
+		JWT:       NewJWTConfig(),
 	}
 }
 
 // Sign method to sign an HTTP request for configured auth strategies.
 func (c WebsocketConfig) Sign(req *http.Request) error {
 	if err := c.OAuth.Sign(req); err != nil {
+		return err
+	}
+	if err := c.JWT.Sign(req); err != nil {
 		return err
 	}
 	return c.BasicAuth.Sign(req)